+++ src/main/java/com/sjodin/thesis/optimisation/GradientDescent.java
@@ -1,0 +1,14 @@
+package com.sjodin.thesis.optimisation;
+
+import com.sjodin.thesis.statements.StatementTree;
+
+// An interface for Gradient Descent over some kind of number (dual numbers, generally)
+// Can find minimum or maximum, as well as giving a program to optimise
+public interface GradientDescent<T> {
+
+    T findMinimum(double precision);
+
+    T findMaximum(double precision);
+
+    void giveProgram(StatementTree program, int parameters, Double[] startPoint);
+}
